// task-manager-client/src/app/models/task.model.ts
export enum TaskStatus {
    OPEN = "OPEN",
    IN_PROGRESS = "IN_PROGRESS",
    DONE = "DONE",
}

export interface Task {
    id: string
    title: string
    description?: string
    status: TaskStatus
    createdAt: Date
    updatedAt?: Date
}

export interface CreateTaskRequest {
    title: string
    description?: string
}

export interface UpdateTaskStatusRequest {
    status: TaskStatus
}